Fix item 2 of Issue #601.
The pointer to the TimeVD StepPointMC collection is given a value in
Mu2eG4/src/Mu2eG4EventAction.cc and is passed into Mu2eG4/src/Mu2eG4SteppingAction.cc,
where it is ultimately used.  When the job is configured not to make a TimeVD
StepPointMC Collection, the pointer has a value of nullptr and the code that would
use it is never called.

The problem was that the call handing the pointer into Mu2eG4SteppingAction turned
the pointer into a reference and then back into a pointer.  That triggered the
sanitize warning.

The solution was to change the signature to always keep it as a pointer.